requirements.txt
ipython
numpy
PyYAML
gnupg
opencv-python
stable-baselines3
hydra-core
wandb
tqdm
scipy
plotly
SciencePlots
pycryptodomex